feat(astro-kbve): mount rn McView with staff gate; remove legacy mc dashboard components
